Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

langchain==0.2.17
langchain-community==0.2.17
langchain-openai

openai==1.30.1

faiss-cpu==1.8.0
pypdf==4.2.0
python-multipart==0.0.9
python-dotenv

pydantic==1.10.13
httpx==0.27.0
pillow==10.4.0
